fix: Fix circular import error due to ConfigManager split from config.py
Due to refactor of our configuration handling ConfigManager was also moved under config/. For backward compatiblity reasons I re-imported it in config.__init__.py but that led to circular imports in specific situations.
