--- src/robocop/config/__init__.py
@@ -1,4 +1,0 @@
-from robocop.config.manager import ConfigManager
-from robocop.config.schema import Config
-
-__all__ = ["Config", "ConfigManager"]
