+++ src/robocop/linter/reports/gitlab.py
@@ -9,7 +9,7 @@ from robocop.files import get_relative_path
 from robocop.linter.rules import RuleSeverity
 
 if TYPE_CHECKING:
-    from robocop.config import Config
+    from robocop.config.schema import Config
     from robocop.linter.diagnostics import Diagnostic, Diagnostics
