+++ src/robocop/linter/reports/json_report.py
@@ -7,7 +7,7 @@ import robocop.linter.reports
 from robocop.files import get_relative_path
 
 if TYPE_CHECKING:
-    from robocop.config import Config
+    from robocop.config.schema import Config
     from robocop.linter.diagnostics import Diagnostic, Diagnostics
