+++ src/robocop/linter/reports/sonarqube.py
@@ -11,8 +11,8 @@ from robocop.linter.rules import RuleSeverity
 if TYPE_CHECKING:
     from pathlib import Path
 
-    from robocop.config import Config
     from robocop.config.manager import ConfigManager
+    from robocop.config.schema import Config
     from robocop.linter.diagnostics import Diagnostic, Diagnostics
     from robocop.linter.rules import Rule
