+++ src/robocop/linter/reports/time_taken_report.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 import robocop.linter.reports
 
 if TYPE_CHECKING:
-    from robocop.config import Config
+    from robocop.config.schema import Config
 
 
 class TimeTakenReport(robocop.linter.reports.ComparableReport):
